ai/requirements.txt
Flask==2.0.3
numpy==1.20.3
opencv_python==4.5.5.62
Pillow==9.0.1
requests==2.26.0
scikit_image==0.18.3

